move unshared input components for entry
